requirements.txt
Flask
Gunicorn
pandas
Flask-Bootstrap==3.3.7.1
numpy
opencv-python==4.7.0.68
requests==2.28.0
pillow
imageio
pybase64
